microservicio-3/requirements.txt
django==4.1.6
django_rest_framework==0.1.0